mgr: collapse rest code into one module
The old multiple modules related to separate services
in Calamari.
