+++ src/pybind/mgr/calamari_rest/manager/osd_request_factory.py
@@ -1,6 +1,6 @@
-from cthulhu.manager.request_factory import RequestFactory
-from calamari_common.types import OsdMap, OSD_IMPLEMENTED_COMMANDS, OSD_FLAGS
-from cthulhu.manager.user_request import OsdMapModifyingRequest, RadosRequest
+from calamari_rest.manager.request_factory import RequestFactory
+from calamari_rest.types import OsdMap, OSD_IMPLEMENTED_COMMANDS, OSD_FLAGS
+from calamari_rest.manager.user_request import OsdMapModifyingRequest, RadosRequest
 from mgr_data import get_sync_object
