+++ src/pybind/mgr/calamari_rest/util.py
@@ -26,3 +26,18 @@ class Ticker(gevent.greenlet.Greenlet):
         while not self._complete.is_set():
             self._callback()
             self._complete.wait(self._period)
+
+
+def memoize(function):
+    def wrapper(*args):
+        self = args[0]
+        if not hasattr(self, "_memo"):
+            self._memo = {}
+
+        if args in self._memo:
+            return self._memo[args]
+        else:
+            rv = function(*args)
+            self._memo[args] = rv
+            return rv
+    return wrapper
