--- src/pybind/mgr/cthulhu/manager/__init__.py
@@ -1,5 +1,0 @@
-
-from calamari_common.config import CalamariConfig
-
-# A config instance for use from within the manager service
-config = CalamariConfig()
